+++ spec/spec_helper.rb
@@ -9,7 +9,7 @@ require "simplecov"
 
 SimpleCov.start do
   enable_coverage :branch
-  add_filter "/spec/"
+  skip "/spec/"
   minimum_coverage line: 95, branch: 90 if ENV["COVERAGE"] == "true"
 end